Added logging again and better error handling
Logging is handled in the main program this time.  Cameras have access
to the log, so they can log events.  Improvements were made for situations
where the tag name doesn't exist and if the cameras were misconfigured.
It should close a little more gracefully and not hang.  It can take some
time to end if a camera is not found.