Added write video path to PLC in config
I have a case where I need the latest video paths to be written to string
tags in the PLC.  If you provide tags to video_path_tags[] in the config,
the paths to the latest videos will be writtin to the tags you provide. Leave
the list empty if you don't want to write, which is default behavior.